Fixed cognito lab integration
Previously the app-api and rest-api-stack were seperate and therefore
the additional endpoints added to the rest-api-stack weren't reachable,
now the app-api code has been merged into rest-api-stack so everything
works together.
