+++ package.json
@@ -27,6 +27,7 @@
     "@aws-sdk/util-dynamodb": "^3.303.0",
     "ajv": "^8.12.0",
     "aws-cdk-lib": "2.88.0",
+    "aws-sdk": "^2.1502.0",
     "axios": "^1.6.2",
     "constructs": "^10.0.0",
     "jsonwebtoken": "^9.0.2",
